fix: maintain single source of DataTypes
