--- frontend/src/container/LogDetailedView/types.ts
@@ -1,10 +1,0 @@
-export enum DataTypes {
-	Int64 = 'int64',
-	String = 'string',
-	Float64 = 'float64',
-	bool = 'bool',
-	ArrayFloat64 = 'array(float64)',
-	ArrayInt64 = 'array(int64)',
-	ArrayString = 'array(string)',
-	ArrayBool = 'array(bool)',
-}
