+++ frontend/src/container/LogsExplorerContext/utils.ts
@@ -1,6 +1,6 @@
 import { OPERATORS } from 'constants/queryBuilder';
-import { DataType } from 'container/MetricsApplication/constant';
 import { ILog } from 'types/api/logs/log';
+import { DataTypes } from 'types/api/queryBuilder/queryAutocompleteResponse';
 import { TagFilterItem } from 'types/api/queryBuilder/queryBuilderData';
 import { v4 as uuid } from 'uuid';
 
@@ -13,7 +13,7 @@ export const getFiltersFromResources = (
 			id: uuid(),
 			key: {
 				key,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				type: 'resource',
 				isColumn: false,
 			},
