+++ frontend/src/container/MetricsApplication/MetricsPageQueries/OverviewQueries.ts
@@ -1,5 +1,8 @@
 import { OPERATORS } from 'constants/queryBuilder';
-import { BaseAutocompleteData } from 'types/api/queryBuilder/queryAutocompleteResponse';
+import {
+	BaseAutocompleteData,
+	DataTypes,
+} from 'types/api/queryBuilder/queryAutocompleteResponse';
 import { TagFilterItem } from 'types/api/queryBuilder/queryBuilderData';
 import {
 	DataSource,
@@ -8,7 +11,6 @@ import {
 } from 'types/common/queryBuilder';
 
 import {
-	DataType,
 	FORMULA,
 	GraphTitle,
 	LATENCY_AGGREGATEOPERATOR,
@@ -40,7 +42,7 @@ export const latency = ({
 		key: isSpanMetricEnable
 			? WidgetKeys.Signoz_latency_bucket
 			: WidgetKeys.DurationNano,
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		type: isSpanMetricEnable ? '' : MetricsType.Tag,
 	};
@@ -52,7 +54,7 @@ export const latency = ({
 			id: '',
 			key: {
 				key: isSpanMetricEnable ? WidgetKeys.Service_name : WidgetKeys.ServiceName,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				type: isSpanMetricEnable ? MetricsType.Resource : MetricsType.Tag,
 				isColumn: !isSpanMetricEnable,
 			},
@@ -62,7 +64,7 @@ export const latency = ({
 		{
 			id: '',
 			key: {
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: !isSpanMetricEnable,
 				key: isSpanMetricEnable ? WidgetKeys.Operation : WidgetKeys.Name,
 				type: MetricsType.Tag,
@@ -98,21 +100,21 @@ export const apDexTracesQueryBuilderQueries = ({
 	threashold,
 }: ApDexProps): QueryBuilderData => {
 	const autoCompleteDataA: BaseAutocompleteData = {
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		key: '',
 		type: '',
 	};
 
 	const autoCompleteDataB: BaseAutocompleteData = {
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		key: '',
 		type: '',
 	};
 
 	const autoCompleteDataC: BaseAutocompleteData = {
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		key: '',
 		type: '',
@@ -123,7 +125,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.ServiceName,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -134,7 +136,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -149,7 +151,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.HasError,
-				dataType: DataType.BOOL,
+				dataType: DataTypes.bool,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -160,7 +162,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.DurationNano,
-				dataType: DataType.FLOAT64,
+				dataType: DataTypes.Float64,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -171,7 +173,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.ServiceName,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -182,7 +184,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -196,7 +198,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.DurationNano,
-				dataType: DataType.FLOAT64,
+				dataType: DataTypes.Float64,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -207,7 +209,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.HasError,
-				dataType: DataType.BOOL,
+				dataType: DataTypes.bool,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -218,7 +220,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.ServiceName,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -229,7 +231,7 @@ export const apDexTracesQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: true,
 				type: MetricsType.Tag,
 			},
@@ -277,21 +279,21 @@ export const apDexMetricsQueryBuilderQueries = ({
 }: ApDexMetricsQueryBuilderQueriesProps): QueryBuilderData => {
 	const autoCompleteDataA: BaseAutocompleteData = {
 		key: WidgetKeys.SignozLatencyCount,
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		type: '',
 	};
 
 	const autoCompleteDataB: BaseAutocompleteData = {
 		key: WidgetKeys.Signoz_latency_bucket,
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		type: '',
 	};
 
 	const autoCompleteDataC: BaseAutocompleteData = {
 		key: WidgetKeys.Signoz_latency_bucket,
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		type: '',
 	};
@@ -301,7 +303,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Service_name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -312,7 +314,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Operation,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -327,7 +329,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.StatusCode,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -338,7 +340,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Le,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -349,7 +351,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Service_name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -360,7 +362,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Operation,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -375,7 +377,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Le,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -386,7 +388,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.StatusCode,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -397,7 +399,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Service_name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -408,7 +410,7 @@ export const apDexMetricsQueryBuilderQueries = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Operation,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -458,7 +460,7 @@ export const operationPerSec = ({
 	const autocompleteData: BaseAutocompleteData[] = [
 		{
 			key: WidgetKeys.SignozLatencyCount,
-			dataType: DataType.FLOAT64,
+			dataType: DataTypes.Float64,
 			isColumn: true,
 			type: '',
 		},
@@ -470,7 +472,7 @@ export const operationPerSec = ({
 				id: '',
 				key: {
 					key: WidgetKeys.Service_name,
-					dataType: DataType.STRING,
+					dataType: DataTypes.String,
 					isColumn: false,
 					type: MetricsType.Resource,
 				},
@@ -481,7 +483,7 @@ export const operationPerSec = ({
 				id: '',
 				key: {
 					key: WidgetKeys.Operation,
-					dataType: DataType.STRING,
+					dataType: DataTypes.String,
 					isColumn: false,
 					type: MetricsType.Tag,
 				},
@@ -510,13 +512,13 @@ export const errorPercentage = ({
 }: OperationPerSecProps): QueryBuilderData => {
 	const autocompleteDataA: BaseAutocompleteData = {
 		key: WidgetKeys.SignozCallsTotal,
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		type: '',
 	};
 	const autocompleteDataB: BaseAutocompleteData = {
 		key: WidgetKeys.SignozCallsTotal,
-		dataType: DataType.FLOAT64,
+		dataType: DataTypes.Float64,
 		isColumn: true,
 		type: '',
 	};
@@ -528,7 +530,7 @@ export const errorPercentage = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Service_name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Resource,
 			},
@@ -539,7 +541,7 @@ export const errorPercentage = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Operation,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -550,7 +552,7 @@ export const errorPercentage = ({
 			id: '',
 			key: {
 				key: WidgetKeys.StatusCode,
-				dataType: DataType.INT64,
+				dataType: DataTypes.Int64,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
@@ -565,7 +567,7 @@ export const errorPercentage = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Service_name,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Resource,
 			},
@@ -576,7 +578,7 @@ export const errorPercentage = ({
 			id: '',
 			key: {
 				key: WidgetKeys.Operation,
-				dataType: DataType.STRING,
+				dataType: DataTypes.String,
 				isColumn: false,
 				type: MetricsType.Tag,
 			},
