+++ frontend/src/container/OnboardingContainer/LogsManagement/common/LogsConnectionStatus/LogsConnectionStatus.tsx
@@ -12,6 +12,7 @@ import { useEffect, useState } from 'react';
 import { SuccessResponse } from 'types/api';
 import { ILog } from 'types/api/logs/log';
 import { MetricRangePayloadProps } from 'types/api/metrics/getQueryRange';
+import { DataTypes } from 'types/api/queryBuilder/queryAutocompleteResponse';
 import { Query } from 'types/api/queryBuilder/queryBuilderData';
 import { EQueryType } from 'types/common/dashboard';
 import { DataSource } from 'types/common/queryBuilder';
@@ -43,7 +44,7 @@ export default function LogsConnectionStatus({
 					aggregateOperator: 'noop',
 					aggregateAttribute: {
 						id: '------false',
-						dataType: '',
+						dataType: DataTypes.EMPTY,
 						key: '',
 						isColumn: false,
 						type: '',
